+++ .storybook/preview-body.html
@@ -1,5 +1,5 @@
 <style>
-  .sb-main-centered #root-inner {
+  /* .sb-main-centered #root-inner {
     width: 1000px;
     height: 400px;
     display: flex;
@@ -8,7 +8,7 @@
   }
   .sb-main-fullscreen #root-inner {
     height: 600px;
-  }
+  } */
   .stacked-form {
     display: flex;
     flex-direction: column;
@@ -19,4 +19,18 @@
   .stacked-form * {
     width: auto;
   }
+
+  .sbdocs * {
+    font-family: RubikVariable, sans-serif;
+  }
+
+  .sbdocs > bl-alert {
+    margin: var(--bl-size-m) 0;
+  }
+
+  .sbdocs [scale] {
+    -webkit-transform: none;
+    transform: none;
+  }
+
 </style>
