+++ README.md
@@ -2,8 +2,23 @@
 
 <h1 align="center">Baklava Design System</h1>
 
-
-[![npm package stable](https://img.shields.io/npm/v/@trendyol/baklava.svg)](https://www.npmjs.com/package/@trendyol/baklava) [![npm package beta](https://img.shields.io/npm/v/@trendyol/baklava/beta.svg)](https://www.npmjs.com/package/@trendyol/baklava/v/beta) [![License](https://img.shields.io/github/license/trendyol/baklava)](https://github.com/Trendyol/baklava/blob/next/LICENSE) [![jsDelivr hits (npm scoped)](https://img.shields.io/jsdelivr/npm/hm/@trendyol/baklava)](https://www.jsdelivr.com/package/npm/@trendyol/baklava) [![semantic-release: angular](https://img.shields.io/badge/semantic--release-angular-e10079?logo=semantic-release)](https://github.com/semantic-release/semantic-release)
+<p align="center">
+  <a href="https://www.npmjs.com/package/@trendyol/baklava">
+    <img src="https://img.shields.io/npm/v/@trendyol/baklava.svg" alt="NPM Package Stable" />
+  </a>
+  <a href="https://www.npmjs.com/package/@trendyol/baklava/v/beta">
+    <img src="https://img.shields.io/npm/v/@trendyol/baklava/beta.svg" alt="NPM Package Beta" />
+  </a>
+  <a href="https://github.com/Trendyol/baklava/blob/next/LICENSE">
+    <img src="https://img.shields.io/github/license/trendyol/baklava" alt="License" />
+  </a>
+  <a href="https://www.jsdelivr.com/package/npm/@trendyol/baklava">
+    <img src="https://img.shields.io/jsdelivr/npm/hm/@trendyol/baklava" alt="jsDelivr hits" />
+  </a>
+  <a href="https://github.com/semantic-release/semantic-release">
+    <img src="https://img.shields.io/badge/semantic--release-angular-e10079?logo=semantic-release" alt="semantic-release: angular" />
+  </a>
+</p>
 
 Baklava is a design system provided by [Trendyol](https://github.com/trendyol) to create a consistent UI/UX for app users.
