+++ docs/customizing-baklava-theme.stories.mdx
@@ -1,16 +1,6 @@
-import { Meta } from '@storybook/addon-docs';
+import { Meta } from '@storybook/blocks';
 
-<Meta
-  title="Documentation/Customizing Baklava Theme"
-  parameters={{
-    viewMode: 'docs',
-    previewTabs: {
-      canvas: {
-        hidden: true,
-      },
-    },
-  }}
-/>
+<Meta title="Documentation/Customizing Baklava Theme" />
 
 # Customizing Baklava Theme
