+++ docs/design-system/adding-a-new-icon.stories.mdx
@@ -1,4 +1,4 @@
-import { Meta } from '@storybook/addon-docs';
+import { Meta } from '@storybook/blocks';
 import createBranchImage from '../images/create-branch-in-vscode.png';
 import iconsFolderImage from '../images/icons-folder.png';
 import stageFilesImage from '../images/stage-files.png';
@@ -7,17 +7,7 @@ import commitPushImage from '../images/commit-and-push.png';
 import openPrImage from '../images/open-pr.png';
 import prFormImage from '../images/pr-form.png';
 
-<Meta
-  title="Design System/Icons/Adding New Icons"
-  parameters={{
-    viewMode: 'docs',
-    previewTabs: {
-      canvas: {
-        hidden: true,
-      },
-    },
-  }}
-/>
+<Meta title="Design System/Icons/Adding New Icons" />
 
 # Adding New Icons to Baklava (For Designers)
