+++ docs/design-system/colors.stories.mdx
@@ -1,4 +1,4 @@
-import { Meta, ColorPalette, ColorItem } from '@storybook/addon-docs';
+import { Meta, ColorPalette, ColorItem } from '@storybook/blocks';
 import { html } from 'lit';
 
 <Meta title="Design System/Colors" />
