+++ docs/design-system/typography.stories.mdx
@@ -1,12 +1,7 @@
-import { Meta, Typeset, Canvas, Story } from '@storybook/addon-docs';
+import { Meta, Typeset, Canvas, Story } from '@storybook/blocks';
 import { html } from 'lit';
 
-<Meta
-  title="Design System/Typography"
-  parameters={{
-    viewMode: 'docs',
-  }}
-/>
+<Meta title="Design System/Typography" />
 
 # Baklava Typography
