+++ package.json
@@ -28,10 +28,10 @@
     "analyze": "cem analyze",
     "commit": "commit",
     "build": "del-cli dist/ && npm run analyze && npm run generate-react-exports && tsc --emitDeclarationOnly && node scripts/build.js",
-    "build-storybook": "NODE_ENV=production build-storybook -o storybook-static",
-    "build-storybook-docs": "build-storybook --docs",
+    "build-storybook": "NODE_ENV=production storybook build -o storybook-static",
+    "build-storybook-docs": "storybook build --docs",
     "serve": "node scripts/build.js --serve",
-    "storybook:dev": "start-storybook -p 1994",
+    "storybook:dev": "storybook dev -p 1994",
     "lint": "npm-run-all -s lint:*",
     "lint:tsc": "tsc --noEmit",
     "lint:eslint": "eslint .",
@@ -79,16 +79,18 @@
     "@custom-elements-manifest/analyzer": "^0.6.3",
     "@open-wc/testing": "^3.1.6",
     "@rollup/plugin-replace": "^4.0.0",
-    "@storybook/addon-a11y": "6.5.16",
-    "@storybook/addon-actions": "6.5.16",
-    "@storybook/addon-essentials": "6.5.16",
-    "@storybook/addon-interactions": "^6.5.16",
-    "@storybook/addon-links": "6.5.16",
-    "@storybook/addon-storysource": "6.5.16",
-    "@storybook/addon-viewport": "^6.5.16",
-    "@storybook/jest": "^0.0.10",
-    "@storybook/testing-library": "^0.0.13",
-    "@storybook/web-components": "6.5.16",
+    "@storybook/addon-a11y": "7.0.6",
+    "@storybook/addon-actions": "7.0.6",
+    "@storybook/addon-essentials": "7.0.6",
+    "@storybook/addon-interactions": "^7.0.6",
+    "@storybook/addon-links": "7.0.6",
+    "@storybook/addon-mdx-gfm": "^7.0.6",
+    "@storybook/addon-storysource": "7.0.6",
+    "@storybook/addon-viewport": "^7.0.6",
+    "@storybook/jest": "^0.1.0",
+    "@storybook/testing-library": "^0.1.0",
+    "@storybook/web-components": "7.0.6",
+    "@storybook/web-components-vite": "^7.0.6",
     "@types/estree": "0.0.47",
     "@types/prettier": "^2.6.4",
     "@typescript-eslint/eslint-plugin": "^5.18.0",
@@ -115,8 +117,10 @@
     "npm-run-all": "^4.1.5",
     "pascal-case": "^3.1.2",
     "prettier": "^2.0.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rollup-plugin-lit-css": "^4.0.0",
-    "storybook": "6.5.16",
+    "storybook": "7.0.6",
     "stylelint": "^14.11.0",
     "stylelint-config-prettier": "^9.0.3",
     "stylelint-config-standard": "^28.0.0",
