+++ src/components/alert/bl-alert.stories.mdx
@@ -4,7 +4,7 @@ import { styleMap } from 'lit/directives/style-map.js';
 import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 import { INITIAL_VIEWPORTS } from '@storybook/addon-viewport';
 
-import { Meta, Canvas, ArgsTable, Story, Preview, Source } from '@storybook/addon-docs';
+import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs';
 
 <Meta
   title="Components/Alert"
