+++ src/components/checkbox-group/bl-checkbox-group.stories.mdx
@@ -1,8 +1,9 @@
 import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { styleMap } from 'lit/directives/style-map.js';
-import { Meta, Canvas, ArgsTable, Story, Preview, Source } from '@storybook/addon-docs';
+import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs';
 import { userEvent } from '@storybook/testing-library';
+import { extraPadding } from '../../utilities/chromatic-decorators';
 
 <Meta
   title="Components/Checkbox/Checkbox Group"
@@ -66,9 +67,7 @@ If you set a list of `values`, options with those values will be selected.
 Checkbox Group component handles keyboard navigation and highlights active checkbox option while navigating with keyboard. First `Tab` focuses on first available option and user can navigate with arrow keys or `Tab`, `Shift+Tab` within options, and `Space` key for selecting it.
 
 <Canvas>
-  <Story name="Focused option" parameters={{
-    extraPaddingForChromatic: true,
-  }} args={{ label: 'Favorite animals', name: 'favoriteAnimals',  options: [{ value: 'cat', label: 'Cat'}, {value: 'dog', label: 'Dog'}] }} play={focusSecondOption}>
+  <Story name="Focused option" decorators={[ extraPadding ]} args={{ label: 'Favorite animals', name: 'favoriteAnimals',  options: [{ value: 'cat', label: 'Cat'}, {value: 'dog', label: 'Dog'}] }} play={focusSecondOption}>
     {CheckboxGroupTemplate.bind({})}
   </Story>
 </Canvas>
