+++ src/components/dialog/bl-dialog.stories.mdx
@@ -2,8 +2,9 @@ import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { styleMap } from 'lit/directives/style-map.js';
 import { unsafeHTML } from 'lit/directives/unsafe-html.js';
-import { Meta, Canvas, ArgsTable, Story, Preview, Source } from '@storybook/addon-docs';
+import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs';
 import { userEvent } from '@storybook/testing-library';
+import { fullscreenLayout } from '../../utilities/chromatic-decorators';
 
 <Meta
   title="Components/Dialog"
@@ -12,6 +13,9 @@ import { userEvent } from '@storybook/testing-library';
     layout: 'fullscreen',
     chromatic: { viewports: [1000] },
   }}
+  decorators={[
+    fullscreenLayout,
+  ]}
   argTypes={{
     open: {
       control: "boolean",
