+++ src/components/drawer/bl-drawer.stories.mdx
@@ -1,17 +1,22 @@
 import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
-import { Meta, Canvas, ArgsTable, Story, Preview, Source } from '@storybook/addon-docs';
+import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs';
+import { fullscreenLayout, withNoAnimation } from '../../utilities/chromatic-decorators';
+import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 
 <Meta
   title="Components/Drawer"
   component="bl-drawer"
   parameters={{
     layout: 'fullscreen',
-    noAnimationOnChromaticLayout: true,
     chromatic: {
       viewports: [768, 320],
     },
   }}
+  decorators={[
+    fullscreenLayout,
+    withNoAnimation,
+  ]}
   argTypes={{
     open: {
       control: "boolean",
@@ -36,7 +41,7 @@ export const openDialog = async (event, args) => {
   document.getElementById(args.id).open=true
 }
 
-export const DummyContent = () => html`
+export const DummyContent = () => `
   <p>
     Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.
   </p>
@@ -52,7 +57,7 @@ export const DummyContent = () => html`
   <p>
     Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.
   </p>
-`
+`;
 
 export const DrawerTemplate = (args) => html`
   <bl-drawer id=${ifDefined(args.id)}
@@ -60,15 +65,15 @@ export const DrawerTemplate = (args) => html`
              embed-url=${ifDefined(args.embedUrl)}
              external-link=${ifDefined(args.externalLink)}>
     <div style="font: var(--bl-font-body-text-2)">
-      ${ifDefined(args.content)}
+      ${unsafeHTML(args.content)}
     </div>
   </bl-drawer>
-`
+`;
 
 export const StoryTemplate = (args) => html`
   ${DrawerTemplate(args)}
   <bl-button @click="${(event)=>openDialog(event, args)}">${ifDefined(args.buttonText)}</bl-button>
-`
+`;
 
 # Drawer
 
@@ -92,10 +97,10 @@ A drawer presents context-specific information and/or actions without leaving th
 
 ## Usage
 
-<Canvas>
+<Canvas isColumn>
   <Story name="Default Values"
          play={(event) => openDialog(event,{id:'drawer-1'})}
-         args={{id:"drawer-1", buttonText:"default drawer", content:DummyContent()}}>
+         args={{id:"drawer-1", buttonText:"default drawer", content: DummyContent()}}>
     {StoryTemplate.bind({})}
   </Story>
   <Story name="With caption"
