+++ src/components/popover/bl-popover.stories.mdx
@@ -2,14 +2,17 @@ import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs';
 import { userEvent } from '@storybook/testing-library';
+import { centeredLayout } from '../../utilities/chromatic-decorators';
 
 <Meta
   title="Components/Popover"
   component="bl-popover"
   parameters={{
-    layout: 'centered',
     chromatic: { viewports: [1000] },
   }}
+  decorators={[
+    centeredLayout,
+  ]}
   argTypes={{
     placement: {
       control: 'text',
