+++ src/components/tab-group/tab/bl-tab.stories.mdx
@@ -2,7 +2,7 @@ import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { styleMap } from 'lit/directives/style-map.js';
 import { unsafeHTML } from 'lit/directives/unsafe-html.js';
-import { Meta, Canvas, ArgsTable, Story, Preview, Source } from '@storybook/addon-docs';
+import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs';
 import { userEvent } from '@storybook/testing-library';
 
 <Meta
@@ -78,7 +78,7 @@ Within tab groups, tabs are used to represent and activate tab panels. A tab can
 Title can be clarified with caption attribute. Caption is located just below the tab title
 
 <Canvas>
-<Story name="Tab With Caption"  args={{ caption: 'Caption' }}>>
+<Story name="Tab With Caption"  args={{ caption: 'Caption' }}>
     {TabsAttributeTemplate.bind({})}
   </Story>
 </Canvas>
@@ -87,7 +87,7 @@ Title can be clarified with caption attribute. Caption is located just below the
 You can give extra information to user with help-text attribute. help-text attribute adds a `bl-tooltip` component at right of the tab
 
 <Canvas>
-<Story name="Tab With Help Text"  args={{helpText:'helpText' }}>>
+<Story name="Tab With Help Text"  args={{helpText:'helpText' }}>
     {TabsAttributeTemplate.bind({})}
   </Story>
 </Canvas>
@@ -96,7 +96,7 @@ You can give extra information to user with help-text attribute. help-text attri
 A Badge can be added to tab. This Badge is showed with `bl-badge` component internally
 
 <Canvas>
-<Story name="Tab With Badge"  args={{badge:'New' }}>>
+<Story name="Tab With Badge"  args={{badge:'New' }}>
     {TabsAttributeTemplate.bind({})}
   </Story>
 </Canvas>
@@ -105,7 +105,7 @@ A Badge can be added to tab. This Badge is showed with `bl-badge` component inte
 An icon can be added to tab. This icon is showed with `bl-icon` component internally and it's color synced with tab's title color
 
 <Canvas>
-<Story name="Tab With Icon" args={{icon1:'heart', icon2:'settings', icon3:'book', icon4:'academy', icon5:'account', icon6:'attach', icon7:'award', icon8:'browser' }}>>
+<Story name="Tab With Icon" args={{icon1:'heart', icon2:'settings', icon3:'book', icon4:'academy', icon5:'account', icon6:'attach', icon7:'award', icon8:'browser' }}>
     {IconTemplate.bind({})}
  </Story>
 </Canvas>
@@ -114,7 +114,7 @@ An icon can be added to tab. This icon is showed with `bl-icon` component intern
 A notification dot can be used in tab compnenet in order for user to get notified by updates in related content
 
 <Canvas>
-<Story name="Tab With Notification Dot">>
+<Story name="Tab With Notification Dot">
     {NotifyTemplate.bind({})}
   </Story>
 </Canvas>
