+++ src/components/textarea/bl-textarea.stories.mdx
@@ -1,14 +1,15 @@
 import { Meta, Canvas, ArgsTable, Story } from "@storybook/addon-docs";
 import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
+import { extraPadding } from '../../utilities/chromatic-decorators';
 
 
 <Meta
   title="Components/Textarea"
   component="bl-textarea"
-  parameters={{
-    extraPaddingForChromatic: true,
-  }}
+  decorators={[
+    extraPadding,
+  ]}
   argTypes={{
     label: {
       control: 'text',
